// packages/core/src/form/types.ts
export type FormControlValue = string | number | boolean | null | readonly string[];
// ...
export type FormControlApi = {
  controlId: () => string;
  descriptionId: () => string;
  errorMessageId: () => string;
  labelId: () => string;
  name: () => string | undefined;
  value: () => FormControlValue | undefined;
  disabled: () => boolean;
  dirty: () => boolean;
  filled: () => boolean;
  form: () => string | undefined;
  focused: () => boolean;
  invalid: () => boolean;
  readonly: () => boolean;
  required: () => boolean;
  touched: () => boolean;
  validating: () => boolean;
  describedBy: () => string | undefined;
  getControlProps: <T extends HTMLElement = HTMLElement>(
    props?: JSX.HTMLAttributes<T>,
  ) => JSX.HTMLAttributes<T>;
  getRootProps: <T extends HTMLElement = HTMLElement>(
    props?: JSX.HTMLAttributes<T>,
  ) => JSX.HTMLAttributes<T>;
  getLabelProps: <T extends HTMLElement = HTMLLabelElement>(
    props?: JSX.LabelHTMLAttributes<T>,
  ) => JSX.LabelHTMLAttributes<T>;
  getDescriptionProps: <T extends HTMLElement = HTMLElement>(
    props?: JSX.HTMLAttributes<T>,
  ) => JSX.HTMLAttributes<T>;
  getErrorMessageProps: <T extends HTMLElement = HTMLElement>(
    props?: JSX.HTMLAttributes<T>,
  ) => JSX.HTMLAttributes<T>;
  getHiddenInputProps: (
    props?: JSX.InputHTMLAttributes<HTMLInputElement>,
  ) => JSX.InputHTMLAttributes<HTMLInputElement>;
  hiddenInputDescriptors: () => readonly HiddenInputDescriptor[];
  registerDescription: (id?: Accessor<string | undefined>) => () => void;
  registerErrorMessage: (id?: Accessor<string | undefined>) => () => void;
  registerFormReset: (element: Accessor<HTMLElement | undefined>) => void;
  registerFormValueSync: (
    element: Accessor<HTMLInputElement | undefined>,
    onValueChange: (value: string) => void,
  ) => void;
};

export type HiddenInputDescriptor = {
  disabled?: boolean;
  form?: string;
  name: string;
  required?: boolean;
  value: string;
};